+++ server/src/driver/EfunVisitor.ts
@@ -23,11 +23,11 @@ import {
     UnionableTypeSpecifierContext,
     ValidIdentifiersContext,
 } from "../parser3/LPCParser";
-import { EfunSymbol } from "../symbols/methodSymbol";
 import { LpcTypes, StructType } from "../types";
 import { COMMENT_CHANNEL } from "../parser3/LPCPreprocessingLexer";
 import { firstEntry, lastEntry } from "../utils";
 import { MethodParameterSymbol } from "../symbols/variableSymbol";
+import { EfunSymbol } from "../symbols/efunSymbol";
 
 type GenericConstructorParameters<T> = ConstructorParameters<
     new (...args: any[]) => T