+++ server/src/symbols/methodSymbol.ts
@@ -3,9 +3,7 @@ import {
     IType,
     ScopedSymbol,
     MethodSymbol as BaseMethodSymbol,
-    ParameterSymbol,
     BaseSymbol,
-    TypeKind,
 } from "antlr4-c3";
 import {
     IKindSymbol,
@@ -43,7 +41,6 @@ import { addDiagnostic } from "./Symbol";
 import { InheritSuperAccessorSymbol } from "./inheritSymbol";
 import { resolveOfTypeSync } from "../backend/symbol-utils";
 import { asStackValue } from "../backend/CallStackUtils";
-import { ensureLpcConfig } from "../backend/LpcConfig";
 import { MethodParameterSymbol } from "./variableSymbol";
 
 export const MAX_CALLDEPTH_SIZE = 25;
@@ -351,191 +348,5 @@ export class ReturnSymbol extends ScopedSymbol implements IEvaluatableSymbol {
             }
         }
         return undefined;
-    }
-}
-
-export class EfunSymbol
-    extends MethodSymbol
-    implements IKindSymbol, IEvaluatableSymbol
-{
-    public constructor(
-        name: string,
-        public returnType?: IType,
-        public functionModifiers: Set<string> = new Set()
-    ) {
-        super(name);
-    }
-    public get kind() {
-        return SymbolKind.Efun;
-    }
-
-    public getParametersSync() {
-        return getSymbolsOfTypeSync(
-            this,
-            MethodParameterSymbol
-        ) as MethodParameterSymbol[];
-    }
-
-    eval(
-        stack: CallStack,
-        args: IEvaluatableSymbol[] = [],
-        callScope?: RootFrame
-    ) {
-        const ownerProgram = (stack.root.symbol as ContextSymbolTable).owner;
-        const { fileHandler } = ownerProgram;
-
-        const argEval = args?.map((a) => {
-            return a?.eval(stack, callScope) as StackValue;
-        });
-
-        const config = ensureLpcConfig();
-
-        // handle special efuns cases
-        switch (this.name) {
-            // NTBLA: put current object on the stack and return that instead of loading a new
-            // instance of the object
-            case "first_inventory":
-            case "next_inventory":
-                return new StackValue(undefined, LpcTypes.objectType, this);
-            case "this_object":
-                return new StackValue(
-                    new ObjectReferenceInfo(
-                        ownerProgram.fileName,
-                        true,
-                        ownerProgram
-                    ),
-                    LpcTypes.objectType,
-                    this
-                );
-            // TODO: this is just a quick hack to get the player object
-            // need a better way to check if references are already loaded.
-            case "this_interactive":
-            case "this_user":
-            case "this_player":
-                const playerCtx = fileHandler.loadReference(
-                    config.files.player,
-                    this
-                );
-
-                return new StackValue(
-                    new ObjectReferenceInfo(
-                        playerCtx?.fileName,
-                        true,
-                        playerCtx
-                    ),
-                    LpcTypes.objectType,
-                    this
-                );
-            case "users":
-                const playerUsersCtx = fileHandler.loadReference(
-                    config.files.player,
-                    this
-                );
-
-                const playerObj = new ObjectReferenceInfo(
-                    playerUsersCtx?.fileName,
-                    true,
-                    playerUsersCtx
-                );
-                return asStackValue(
-                    [playerObj],
-                    LpcTypes.objectArrayType,
-                    this
-                );
-                break;
-            case "explode":
-                const str = argEval[0];
-                const delim = argEval[1];
-                const strVal = str?.value ?? 0;
-                const delimVal = delim?.value;
-
-                if (
-                    str?.type?.name != "string" ||
-                    delim?.type?.name != "string"
-                ) {
-                    return undefined;
-                }
-                if (strVal == 0 || delimVal == 0 || typeof strVal != "string") {
-                    return undefined;
-                }
-
-                return new ArrayStackValue<string>(
-                    (strVal as string)
-                        ?.split(delimVal)
-                        .map((s) => asStackValue(s, LpcTypes.stringType, this)),
-                    LpcTypes.stringArrayType,
-                    this
-                );
-            case "new":
-            case "clone_object":
-                return cloneObjectImpl(stack, argEval, callScope, this);
-            case "previous_object":
-                return previousObjectImpl(stack);
-        }
-
-        return undefined;
-    }
-
-    public allowsMultiArgs() {
-        const prms = this.getParametersSync();
-        if (prms.length === 0) return false;
-        return prms[prms.length - 1].varArgs;
-    }
-}
-
-/** @deprecated */
-export class EfunParamSymbol extends MethodParameterSymbol {
-    constructor(name: string, type: IType, public allowMulti?: boolean) {
-        super(name, type);
-    }
-}
-
-function previousObjectImpl(stack: CallStack) {
-    const filename = stack.peek().filename;
-    let symbol = stack.peek().symbol;
-
-    for (let i = stack.length - 1; i >= 0; i--) {
-        const frame = stack.at(i);
-        if (frame.filename != filename) {
-            symbol = frame.symbol;
-            break;
-        }
-    }
-
-    const ownerCtx = (symbol?.symbolTable as ContextSymbolTable)?.owner;
-    const info = new ObjectReferenceInfo();
-    info.context = ownerCtx;
-    info.filename = ownerCtx?.fileName;
-    info.isLoaded = !!ownerCtx;
-    return asStackValue(info);
-}
-
-function cloneObjectImpl(
-    stack: CallStack,
-    argVals: StackValue[],
-    callScope: RootFrame,
-    symbol: EfunSymbol
-) {
-    const ownerProgram = (stack.root.symbol as ContextSymbolTable).owner;
-    const { fileHandler } = ownerProgram;
-
-    // what type is the first arg?
-    const firstArg = argVals.at(0);
-    if (!firstArg) {
-        return undefined;
-    }
-
-    if (firstArg.type.kind == TypeKind.String) {
-        const filename = firstArg.value;
-        const ctx = fileHandler.loadReference(filename, symbol);
-        if (!ctx) {
-            return undefined;
-        }
-
-        const info = new ObjectReferenceInfo();
-        info.filename = filename;
-        info.isLoaded = true;
-        info.context = ctx;
-        return asStackValue(info);
     }
 }